# docs/openapi-draft.yaml
openapi: 3.0.3
info:
  title: Zoodopt API
  version: 2.0.0
  description: REST API for pet adoption platform connecting shelters with adopters.
  contact:
    name: Zoodopt Development Team
    email: support@zoodopt.com
servers:
  - url: http://localhost:5217
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
        timestamp:
          type: string
          format: date-time
          description: When the error occurred

    # Auth DTOs
...
    ApplicationCreateDto:
      type: object
      required:
        - petId
      properties:
        petId:
          type: integer
          example: 1
        message:
          type: string
          nullable: true
          example: "I would love to adopt this pet!"
    ApplicationMessageDto:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: "Updated message about why I want to adopt"
    ApplicationStatusDto:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum: [Pending, Approved, Rejected]
          example: "Approved"
    ApplicationResponseDto:
      type: object
      properties:
        id:
          type: integer
        petId:
          type: integer
        userId:
          type: integer
        message:
          type: string
        status:
          type: string
        submittedAt:
          type: string
          format: date-time
        petName:
          type: string
          nullable: true

    # Shelter DTOs
...
paths:
...
  /applications:
    get:
      tags:
        - Applications
      summary: List applications
      description: View applications for shelter's pets (shelter staff and admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApplicationResponseDto"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      tags:
        - Applications
      summary: Create application
      description: Submit an adoption application
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationCreateDto"
      responses:
        "201":
          description: Application submitted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationResponseDto"
        "400":
          description: Validation failed or pet doesn't exist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Duplicate application for the same pet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /applications/my-applications:
    get:
      tags:
        - Applications
      summary: Get current user's applications
      description: Get all applications submitted by the current authenticated user, ordered by submission date (newest first)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of user's applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApplicationResponseDto"
              example:
                - id: 4
                  petId: 2
                  userId: 2
                  message: "I'm an experienced adopter."
                  status: "Pending"
                  submittedAt: "2025-07-31T17:09:04.000Z"
                  petName: "Whiskers"
                - id: 1
                  petId: 1
                  userId: 2
                  message: "I love dogs and have experience adopting."
                  status: "Approved"
                  submittedAt: "2025-07-31T05:02:37.000Z"
                  petName: "Max"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /applications/{id}:
    get:
      tags:
        - Applications
      summary: Get application by ID
      description: Get details for a specific application
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        "200":
          description: Application details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationResponseDto"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Application not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      tags:
        - Applications
      summary: Delete application
      description: Delete an application (applicant or admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        "204":
          description: Application deleted successfully
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Application not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /applications/{id}/message:
    put:
      tags:
        - Applications
      summary: Update application message
      description: Update application message (applicant only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationMessageDto"
      responses:
        "200":
          description: Message updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationResponseDto"
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Application not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /applications/{id}/status:
    put:
      tags:
        - Applications
      summary: Update application status
      description: Update application status (shelter staff and admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationStatusDto"
      responses:
        "200":
          description: Status updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationResponseDto"
        "400":
          description: Validation failed or invalid status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Application not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Shelter endpoints
...